Big changes, check description
Transitioned the Simulation to support a new class "Run" that contains a lot of the information that was held in Simulation.

The Run class now allows you to change between different Runs of the same or different algorithms.

Added the the GUI the Vacuum Runs dropdown and that allows the user to control which run is being visualised.

Added the trails to the Run class but there are some SERIOUS performance issues and I do not know if there's a simple fix.

Autopilot works although some of the code was commented out.

Bugfix that didn't reset the time when stopped if the game had already previously ran though a simulation completely one time.

Bugfix that fixed the house planner from displaying a different floor than what was said in the GUI.